+++ packages/react/tailwind.config.js
@@ -2,8 +2,8 @@ import { colors } from '@click-tools-ui/tokens'
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
+  content: ['./src/**/*.{js,ts,jsx,tsx}'],
   theme: {
     colors,
   },
-  content: ['./src/**/*.{js,ts,jsx,tsx}'],
 }